requirements.txt
# Core
langgraph>=0.2.0
langchain>=0.3.0
langchain-ollama>=0.2.0

# External APIs
requests>=2.31.0

# Memory & Storage
chromadb>=0.5.0

# Environment
python-dotenv>=1.0.0

# Data Handling
pydantic>=2.5.0

# UI & API Server
streamlit>=1.35.0
fastapi>=0.110.0
uvicorn>=0.27.0
reportlab>=4.2.0
